Add files via upload
It is  a simple word guessing game in which you have different words and computer would pick a random one and you have to guess the letters of the word. You have 6 lives and when you gave a correct guess of letter it will take you to another step. And if you can't guess the letter then you will lost your one live.